- Documentation.
git-svn-id: http://subversion.assembla.com/svn/hyperion_server@61 c48d7634-943c-464e-a392-e2f9434beeda

+++ src/com/grahamedgecombe/rs2/event/Event.java
@@ -1,33 +1,69 @@
 package com.grahamedgecombe.rs2.event;
 
+/**
+ * Represents a task that is executed in the future, once or periodically.
+ * @author Graham
+ *
+ */
 public abstract class Event {
 	
-	private int delay;
+	/**
+	 * The delay, in milliseconds.
+	 */
+	private long delay;
+	
+	/**
+	 * The running flag.
+	 */
 	private boolean running = true;
 	
-	public Event(int delay) {
+	/**
+	 * Creates an event with the specified delay.
+	 * @param delay The delay.
+	 */
+	public Event(long delay) {
 		this.delay = delay;
 	}
 	
-	public int getDelay() {
+	/**
+	 * Gets the event delay.
+	 * @return The delay, in milliseconds.
+	 */
+	public long getDelay() {
 		return delay;
 	}
 	
-	public void setDelay(int delay) {
+	/**
+	 * Sets the event delay.
+	 * @param delay The delay to set.
+	 * @throws IllegalArgumentException if the delay is negative.
+	 */
+	public void setDelay(long delay) {
 		if(delay < 0) {
 			throw new IllegalArgumentException("Delay must be positive.");
 		}
 		this.delay = delay;
 	}
 	
+	/**
+	 * Checks if the event is running.
+	 * @return <code>true</code> if the event is still running, <code>false</code> if not.
+	 */
 	public boolean isRunning() {
 		return running;
 	}
 	
+	/**
+	 * Stops the event from running in the future.
+	 */
 	public void stop() {
 		running = false;
 	}
 	
+	/**
+	 * The execute method is called when the event is run. The general contract
+	 * of the execute method is that it may take any action whatsoever.
+	 */
 	public abstract void execute();
 
 }
